fix some CyGroupXcorr code (still crashing), add init to cython ext
also made some small warning fixes to ipp_ext, add imports to
xcorrRoutines

+++ cython_ext/ipp_ext.h
@@ -292,7 +292,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp8u *newm_data = ippsMalloc_8u_L(cap);
-			ippsCopy_8u(m_data, newm_data, copylen);
+			ippsCopy_8u(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -307,7 +307,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp16u *newm_data = ippsMalloc_16u_L(cap);
-			ippsCopy_16s((Ipp16s*)m_data, (Ipp16s*)newm_data, copylen); // reuse 16s for non-existent 16u
+			ippsCopy_16s((Ipp16s*)m_data, (Ipp16s*)newm_data, (int)copylen); // reuse 16s for non-existent 16u
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -322,7 +322,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp32u *newm_data = ippsMalloc_32u_L(cap);
-			ippsCopy_32s((Ipp32s*)m_data, (Ipp32s*)newm_data, copylen); // reuse 32s
+			ippsCopy_32s((Ipp32s*)m_data, (Ipp32s*)newm_data, (int)copylen); // reuse 32s
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -337,7 +337,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp8s *newm_data = ippsMalloc_8s_L(cap);
-			ippsCopy_8u((Ipp8u*)m_data, (Ipp8u*)newm_data, copylen); // reuse 8u
+			ippsCopy_8u((Ipp8u*)m_data, (Ipp8u*)newm_data, (int)copylen); // reuse 8u
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -352,7 +352,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp16s *newm_data = ippsMalloc_16s_L(cap);
-			ippsCopy_16s(m_data, newm_data, copylen);
+			ippsCopy_16s(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -367,7 +367,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp32s *newm_data = ippsMalloc_32s_L(cap);
-			ippsCopy_32s(m_data, newm_data, copylen);
+			ippsCopy_32s(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -382,7 +382,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp64s *newm_data = ippsMalloc_64s_L(cap);
-			ippsCopy_64s(m_data, newm_data, copylen);
+			ippsCopy_64s(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -397,7 +397,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp32f *newm_data = ippsMalloc_32f_L(cap);
-			ippsCopy_32f(m_data, newm_data, copylen);
+			ippsCopy_32f(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -412,7 +412,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp64f *newm_data = ippsMalloc_64f_L(cap);
-			ippsCopy_64f(m_data, newm_data, copylen);
+			ippsCopy_64f(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -427,7 +427,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp8sc *newm_data = ippsMalloc_8sc_L(cap);
-			ippsCopy_16s((Ipp16s*)m_data, (Ipp16s*)newm_data, copylen); // reuse 16s
+			ippsCopy_16s((Ipp16s*)m_data, (Ipp16s*)newm_data, (int)copylen); // reuse 16s
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -442,7 +442,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp16sc *newm_data = ippsMalloc_16sc_L(cap);
-			ippsCopy_16sc(m_data, newm_data, copylen);
+			ippsCopy_16sc(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -457,7 +457,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp32sc *newm_data = ippsMalloc_32sc_L(cap);
-			ippsCopy_32sc(m_data, newm_data, copylen);
+			ippsCopy_32sc(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -472,7 +472,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp64sc *newm_data = ippsMalloc_64sc_L(cap);
-			ippsCopy_64sc(m_data, newm_data, copylen);
+			ippsCopy_64sc(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -487,7 +487,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp32fc *newm_data = ippsMalloc_32fc_L(cap);
-			ippsCopy_32fc(m_data, newm_data, copylen);
+			ippsCopy_32fc(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
@@ -502,7 +502,7 @@ namespace ippe
 				
 		if (reMalloc){
 			Ipp64fc *newm_data = ippsMalloc_64fc_L(cap);
-			ippsCopy_64fc(m_data, newm_data, copylen);
+			ippsCopy_64fc(m_data, newm_data, (int)copylen);
 			ippsFree(m_data);
 			m_data = newm_data;
 			check_malloc();
